refactor tests : 1) use Junit4; 2) avoid creating user/login every test case
