+++ android/src/androidTest/java/com/magnet/mmx/client/ClientTestConfigImpl.java
@@ -17,7 +17,6 @@ package com.magnet.mmx.client;
 
 import android.content.Context;
 import android.os.Environment;
-
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
