+++ android/src/androidTest/java/com/magnet/mmx/client/api/ExecMonitor.java
@@ -14,8 +14,10 @@
  */
 package com.magnet.mmx.client.api;
 
-class ExecMonitor<S, F> {
-  enum Status {
+import com.magnet.mmx.client.common.Log;
+
+public class ExecMonitor<S, F> {
+  public enum Status {
     INVOKED,
     FAILED,
     WAITING,
@@ -23,14 +25,23 @@ class ExecMonitor<S, F> {
 
   private S mRtnValue;
   private F mFailedValue;
-  private ExecMonitor.Status mStatus = Status.WAITING;
+  private volatile ExecMonitor.Status mStatus = Status.WAITING;
+  private String mName;
   
   public void reset(S rtnValue, F failValue) {
     mRtnValue = rtnValue;
     mFailedValue = failValue;
     mStatus = Status.WAITING;
   }
-  
+
+  public ExecMonitor(String name) {
+    this.mName = name;
+  }
+
+  public ExecMonitor() {
+    this("");
+  }
+
   // Only be called if waitFor() returns EXECED
   public S getReturnValue() {
     return mRtnValue;
@@ -44,24 +55,30 @@ class ExecMonitor<S, F> {
   public synchronized void invoked(S value) {
     mStatus = Status.INVOKED;
     mRtnValue = value;
-    notify();
+    notifyAll();
 
   }
   
   public synchronized void failed(F value) {
     mStatus = Status.FAILED;
     mFailedValue = value;
-    notify();
+    notifyAll();
   }
   
   public synchronized ExecMonitor.Status waitFor(long timeout) {
+    long startTime = System.currentTimeMillis();
     if (mStatus == Status.WAITING) {
       // Not executed yet, wait for result.
+      Log.d("ExecMonitor", "\n-----------------------------" + mName + " waiting-----------------------------\n");
       try {
         wait(timeout);
+        Log.d("ExecMonitor", "\n-----------------------------" + mName + " waiting done after " + (System.currentTimeMillis() - startTime) + "-----------------------------\n");
       } catch (InterruptedException e) {
-        // Ignored
+        //Log.e("ExecMonitor", "Timeout", e);
+        Log.d("ExecMonitor", "\n-----------------------------" + mName + " waiting timeout after " + (System.currentTimeMillis() - startTime) + "-----------------------------\n");
       }
+    } else {
+      Log.d("ExecMonitor", "\n-----------------------------" + mName + " no need to waiting " + mStatus + "-----------------------------\n");
     }
     return mStatus;
   }
