+++ android/src/androidTest/java/com/magnet/mmx/client/api/MMXMessageTest.java
@@ -14,251 +14,55 @@
  */
 package com.magnet.mmx.client.api;
 
-import com.magnet.max.android.Attachment;
+import android.support.test.runner.AndroidJUnit4;
+import com.magnet.max.android.User;
+import com.magnet.mmx.client.common.Log;
+import com.magnet.mmx.client.utils.MaxHelper;
+import com.magnet.mmx.client.utils.TestCaseTimer;
+import com.magnet.mmx.client.utils.TestConstants;
+import com.magnet.mmx.client.utils.UserHelper;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestName;
+import org.junit.runner.RunWith;
 
-import com.magnet.max.android.User;
-import com.magnet.mmx.client.api.MMXMessage.InvalidRecipientException;
-import com.magnet.mmx.client.common.Log;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.AtomicReference;
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.fail;
 
-public class MMXMessageTest extends MMXInstrumentationTestCase {
+@RunWith(AndroidJUnit4.class)
+public class MMXMessageTest {
   private static final String TAG = MMXMessageTest.class.getSimpleName();
 
-  @Override
-  public void tearDown() {
-    helpLogout();
+  @Rule
+  public TestCaseTimer testCaseTimer = new TestCaseTimer();
+
+  @BeforeClass
+  public static void setUp() {
+    MaxHelper.initMax();
   }
 
+  @AfterClass
+  public static void tearDown() {
+    UserHelper.logout();;
+  }
+
+  @Test
   public void testFailureCodes() {
     assertEquals(MMXMessage.FailureCode.BAD_REQUEST, MMX.FailureCode.BAD_REQUEST);
     assertEquals(MMXMessage.FailureCode.BAD_REQUEST, 
         MMXMessage.FailureCode.fromMMXFailureCode(MMX.FailureCode.BAD_REQUEST, null));
   }
-  
-  public void testSendMessage() {
-    String suffix = String.valueOf(System.currentTimeMillis());
-    String username = USERNAME_PREFIX + suffix;
-    String displayName = DISPLAY_NAME_PREFIX + suffix;
-    registerUser(username, displayName, PASSWORD);
 
-    //login with credentials
-    loginMax(username, new String(PASSWORD));
-
-    assertTrue(MMX.getMMXClient().isConnected());
-    MMX.start();
-    final ExecMonitor<HashMap<String, Object>, Void> receivedResult = new ExecMonitor<HashMap<String, Object>, Void>();
-    final StringBuffer senderBuffer = new StringBuffer();
-    final ExecMonitor<String, Void> acknowledgeResult = new ExecMonitor<String, Void>();
-    MMX.EventListener messageListener = new MMX.EventListener() {
-      public boolean onMessageReceived(MMXMessage message) {
-        Log.d(TAG, "onMessageReceived(): " + message.getId());
-        senderBuffer.append(message.getSender().getFirstName());
-        HashMap<String, Object> receivedContent = new HashMap<String, Object>();
-        for (Map.Entry<String, String> entry : message.getContent().entrySet()) {
-          receivedContent.put(entry.getKey(), entry.getValue());
-        }
-        receivedResult.invoked(receivedContent);
-        //do the acknowledgement
-        message.acknowledge(null);
-        return false;
-      }
-
-      public boolean onMessageAcknowledgementReceived(User from, String messageId) {
-        acknowledgeResult.invoked(messageId);
-        return false;
-      }
-    };
-    MMX.registerListener(messageListener);
-
-    HashSet<User> recipients = new HashSet<User>();
-    recipients.add(MMX.getCurrentUser());
-
-    HashMap<String, String> content = new HashMap<String, String>();
-    content.put("foo", "bar");
-    MMXMessage message = new MMXMessage.Builder()
-            .recipients(recipients)
-            .content(content)
-            .build();
-    final ExecMonitor<String, Boolean> sendResult = new ExecMonitor<String, Boolean>();
-    final String messageId = message.send(new MMXMessage.OnFinishedListener<String>() {
-      public void onSuccess(String result) {
-        Log.e(TAG, "testSendMessage(): onSuccess() msgId=" + result);
-        sendResult.invoked(result);
-      }
-
-      public void onFailure(MMXMessage.FailureCode code, Throwable ex) {
-        Log.e(TAG, "testSendMessage(): failureCode=" + code, ex);
-        sendResult.failed(Boolean.TRUE);
-      }
-    });
-    // Check if the send is success
-    ExecMonitor.Status status = sendResult.waitFor(TIMEOUT_IN_MILISEC);
-    assertEquals(ExecMonitor.Status.INVOKED, status);
-    assertEquals(messageId, sendResult.getReturnValue());
-    
-    // Check if the receive is success
-    status = receivedResult.waitFor(TIMEOUT_IN_MILISEC);
-    if (status == ExecMonitor.Status.WAITING) {
-      fail("testSendMessage() receive msg timed out");
-    }
-    assertEquals("bar", receivedResult.getReturnValue().get("foo"));
-    assertEquals(MMX.getCurrentUser().getFirstName(), senderBuffer.toString());
-
-    //check acknowledgement
-
-    status = acknowledgeResult.waitFor(TIMEOUT_IN_MILISEC);
-    if (status == ExecMonitor.Status.WAITING) {
-      fail("testSenddMessage() receive acknowledgement timed out");
-    }
-    assertEquals(messageId, acknowledgeResult.getReturnValue());
-
-    MMX.unregisterListener(messageListener);
-//    assertFalse(MMX.getMMXClient().isConnected());
-  }
-
-  public void testSendMessageWithAttachments() {
-    String suffix = String.valueOf(System.currentTimeMillis());
-    String username = USERNAME_PREFIX + suffix;
-    String displayName = DISPLAY_NAME_PREFIX + suffix;
-    registerUser(username, displayName, PASSWORD);
-
-    //login with credentials
-    loginMax(username, new String(PASSWORD));
-
-    assertTrue(MMX.getMMXClient().isConnected());
-    MMX.start();
-
-    final AtomicLong attachmentSize = new AtomicLong();
-    final AtomicReference<Attachment> attachmentRef = new AtomicReference<>();
-
-    final ExecMonitor<HashMap<String, Object>, Void> receivedResult = new ExecMonitor<HashMap<String, Object>, Void>();
-    final StringBuffer senderBuffer = new StringBuffer();
-    final ExecMonitor<String, Void> acknowledgeResult = new ExecMonitor<String, Void>();
-    MMX.EventListener messageListener = new MMX.EventListener() {
-      public boolean onMessageReceived(MMXMessage message) {
-        Log.d(TAG, "onMessageReceived(): " + message.getId());
-        senderBuffer.append(message.getSender().getFirstName());
-        HashMap<String, Object> receivedContent = new HashMap<String, Object>();
-        for (Map.Entry<String, String> entry : message.getContent().entrySet()) {
-          receivedContent.put(entry.getKey(), entry.getValue());
-        }
-
-        //Attachments
-        assertNotNull(message.getAttachments());
-        assertEquals(1, message.getAttachments().size());
-        Attachment attachmentReceived = message.getAttachments().get(0);
-        assertEquals("image/jpeg", attachmentReceived.getMimeType());
-        assertEquals(Attachment.Status.INIT, attachmentReceived.getStatus());
-        //assertEquals(attachmentSize.get(), attachmentReceived.getLength());
-        assertNotNull(attachmentReceived.getDownloadUrl());
-
-        attachmentRef.set(attachmentReceived);
-
-        receivedResult.invoked(receivedContent);
-
-        //do the acknowledgement
-        message.acknowledge(null);
-
-        return false;
-      }
-
-      public boolean onMessageAcknowledgementReceived(User from, String messageId) {
-        acknowledgeResult.invoked(messageId);
-        return false;
-      }
-    };
-    MMX.registerListener(messageListener);
-
-    HashSet<User> recipients = new HashSet<User>();
-    recipients.add(MMX.getCurrentUser());
-
-    HashMap<String, String> content = new HashMap<String, String>();
-    content.put("foo", "bar");
-    final Attachment attachment1 = new Attachment(getContext().getResources().openRawResource(
-        com.magnet.mmx.test.R.raw.test_image), "image/jpeg");
-    //final Attachment attachment1 = new TextAttachment(Attachment.TEXT_PLAIN, "hello world");
-    //assertEquals(-1, attachment1.getLength());
-    MMXMessage message = new MMXMessage.Builder()
-        .recipients(recipients)
-        .content(content)
-        .attachments(attachment1)
-        .build();
-    final ExecMonitor<String, Boolean> sendResult = new ExecMonitor<String, Boolean>();
-    final String messageId = message.send(new MMXMessage.OnFinishedListener<String>() {
-      public void onSuccess(String result) {
-        Log.e(TAG, "testSendMessage(): onSuccess() msgId=" + result);
-
-        assertTrue(attachment1.getLength() > 0);
-        attachmentSize.set(attachment1.getLength());
-
-        sendResult.invoked(result);
-      }
-
-      public void onFailure(MMXMessage.FailureCode code, Throwable ex) {
-        Log.e(TAG, "testSendMessage(): failureCode=" + code, ex);
-        sendResult.failed(Boolean.TRUE);
-      }
-    });
-    // Check if the send is success
-    ExecMonitor.Status status = sendResult.waitFor(20 * 1000);
-    assertEquals(ExecMonitor.Status.INVOKED, status);
-    assertEquals(messageId, sendResult.getReturnValue());
-
-    // Check if the receive is success
-    status = receivedResult.waitFor(30 * 1000);
-    if (status == ExecMonitor.Status.WAITING) {
-      fail("testSendMessage() receive msg timed out");
-    }
-    assertEquals("bar", receivedResult.getReturnValue().get("foo"));
-    assertEquals(MMX.getCurrentUser().getFirstName(), senderBuffer.toString());
-
-    //check acknowledgement
-    status = acknowledgeResult.waitFor(TIMEOUT_IN_MILISEC);
-    if (status == ExecMonitor.Status.WAITING) {
-      fail("testSenddMessage() receive acknowledgement timed out");
-    }
-    assertEquals(messageId, acknowledgeResult.getReturnValue());
-
-    // Download attachment
-    assertNotNull(attachmentRef.get());
-    final CountDownLatch downLatch = new CountDownLatch(1);
-    final Attachment attachmentReceived = attachmentRef.get();
-    Log.d(TAG, "-----------attachment received : " + attachmentReceived.getDownloadUrl());
-    attachmentReceived.download(new Attachment.DownloadAsBytesListener() {
-
-      @Override public void onComplete(byte[] bytes) {
-        assertNotNull(bytes);
-        assertTrue(bytes.length > 0);
-        assertEquals(attachmentReceived.getLength(), bytes.length);
-        downLatch.countDown();
-      }
-
-      @Override public void onError(Throwable throwable) {
-        fail(throwable.getMessage());
-      }
-    });
-    try {
-      downLatch.await(TIMEOUT_IN_MILISEC, TimeUnit.MILLISECONDS);
-    } catch (InterruptedException e) {
-      fail(e.getMessage());
-    }
-    assertEquals(0, downLatch.getCount());
-    assertEquals(Attachment.Status.COMPLETE, attachmentReceived.getStatus());
-
-    MMX.unregisterListener(messageListener);
-    //    assertFalse(MMX.getMMXClient().isConnected());
-  }
-  
+  @Test
   public void testSendBeforeLogin() {
     HashSet<User> recipients = new HashSet<User>();
-    recipients.add(getInvalidUser("foo"));
+    recipients.add(UserHelper.getInvalidUser("foo"));
     HashMap<String,String> content = new HashMap<String,String>();
     content.put("foo","bar");
     MMXMessage message = new MMXMessage.Builder()
@@ -278,7 +82,7 @@ public class MMXMessageTest extends MMXInstrumentationTestCase {
         failureMonitor.failed(code);
       }
     });
-    ExecMonitor.Status status = failureMonitor.waitFor(TIMEOUT_IN_MILISEC);
+    ExecMonitor.Status status = failureMonitor.waitFor(TestConstants.TIMEOUT_IN_MILISEC);
     if (status == ExecMonitor.Status.INVOKED) {
       fail("should have called onFailure()");
     } else if (status == ExecMonitor.Status.FAILED) {
@@ -316,179 +120,4 @@ public class MMXMessageTest extends MMXInstrumentationTestCase {
 //      fail("channel.publish() timed out");
 //    }
 //  }
-
-  public void testSendUCastMessageError() {
-    String suffix = String.valueOf(System.currentTimeMillis());
-    String username = USERNAME_PREFIX + suffix;
-    String displayName = DISPLAY_NAME_PREFIX + suffix;
-    String noSuchUser = NO_SUCH_USERNAME_PREFIX;
-    registerUser(username, displayName, PASSWORD);
-    
-    final Set<String> invalidUsers = new HashSet<String>();
-    
-    //login with credentials
-    loginMax(username, new String(PASSWORD));
-
-    assertTrue(MMX.getMMXClient().isConnected());
-    MMX.start();
-    final ExecMonitor<Boolean, MMXMessage.FailureCode> receivedResult =
-        new ExecMonitor<Boolean, MMXMessage.FailureCode>();
-    MMX.EventListener messageListener = new MMX.EventListener() {
-      @Override
-      public boolean onMessageReceived(MMXMessage message) {
-        Log.d(TAG, "onMessageReceived(): " + message.getId());
-        receivedResult.invoked(Boolean.TRUE);
-        return false;
-      }
-      @Override
-      public boolean onMessageAcknowledgementReceived(User from, String messageId) {
-        receivedResult.invoked(Boolean.TRUE);
-        return false;
-      }
-      @Override
-      public boolean onMessageSendError(String messageId, 
-                        MMXMessage.FailureCode code, String text) {
-        Log.d(TAG, "onMessageSendError(): msgId="+messageId+", code="+code+", text="+text);
-        receivedResult.failed(code);
-        return false;
-      }
-    };
-    MMX.registerListener(messageListener);
-
-    User fooUser = getInvalidUser("foo");
-    HashSet<User> recipients = new HashSet<User>();
-    recipients.add(fooUser);
-    
-    HashMap<String, String> content = new HashMap<String, String>();
-    content.put("foo", "bar");
-    MMXMessage message = new MMXMessage.Builder()
-            .recipients(recipients)
-            .content(content)
-            .build();
-    final ExecMonitor<Boolean, MMXMessage.FailureCode> sendResult =
-            new ExecMonitor<Boolean, MMXMessage.FailureCode>();
-    final String messageId = message.send(new MMXMessage.OnFinishedListener<String>() {
-      @Override
-      public void onSuccess(String msgId) {
-        sendResult.invoked(Boolean.TRUE);
-      }
-      @Override
-      public void onFailure(MMXMessage.FailureCode code, Throwable ex) {
-        Log.d(TAG, "Unicast Send Msg failed: code="+code+", ex="+ex);
-        if (code.equals(MMXMessage.FailureCode.INVALID_RECIPIENT)) {
-          InvalidRecipientException irEx = (InvalidRecipientException) ex;
-          invalidUsers.addAll(irEx.getUserIds());
-        }
-        sendResult.failed(code);
-      }
-    });
-    
-    // Send failed because of an invalid recipient
-    ExecMonitor.Status status = sendResult.waitFor(TIMEOUT_IN_MILISEC);
-    assertEquals(ExecMonitor.Status.FAILED, status);
-    assertEquals(MMXMessage.FailureCode.INVALID_RECIPIENT, sendResult.getFailedValue());
-    assertEquals(1, invalidUsers.size());
-    assertTrue(invalidUsers.contains(fooUser.getUserIdentifier()));
-    
-    // Make sure that no error message came and no message arrived.
-    status = receivedResult.waitFor(2000);
-    assertEquals(ExecMonitor.Status.WAITING, status);
-
-    MMX.unregisterListener(messageListener);
-//    assertFalse(MMX.getMMXClient().isConnected());
-  }
-  
-  public void testSendMCastMessageError() {
-    String suffix = String.valueOf(System.currentTimeMillis());
-    String username = USERNAME_PREFIX + suffix;
-    String displayName = DISPLAY_NAME_PREFIX + suffix;
-    String noSuchUser = NO_SUCH_USERNAME_PREFIX;
-    String wrongUser = WRONG_USERNAME_PREFIX;
-    registerUser(username, displayName, PASSWORD);
-
-    final Set<String> invalidUsers = new HashSet<String>();
-    
-    //login with credentials
-    loginMax(username, new String(PASSWORD));
-
-    assertTrue(MMX.getMMXClient().isConnected());
-    MMX.start();
-    final ExecMonitor<Boolean, MMXMessage.FailureCode> receivedResult =
-            new ExecMonitor<Boolean, MMXMessage.FailureCode>();
-    MMX.EventListener messageListener = new MMX.EventListener() {
-      @Override
-      public boolean onMessageReceived(MMXMessage message) {
-        Log.d(TAG, "onMessageReceived(): " + message.getId());
-        receivedResult.invoked(Boolean.TRUE);
-        return false;
-      }
-      @Override
-      public boolean onMessageAcknowledgementReceived(User from, String messageId) {
-        receivedResult.invoked(Boolean.TRUE);
-        return false;
-      }
-      @Override
-      public boolean onMessageSendError(String messageId, 
-                        MMXMessage.FailureCode code, String text) {
-        Log.d(TAG, "onMessageSendError(): msgId="+messageId+", code="+code+", text="+text);
-        receivedResult.failed(code);
-        return false;
-      }
-    };
-    MMX.registerListener(messageListener);
-
-    HashSet<User> recipients = new HashSet<User>();
-    User badRecipient1 = getInvalidUser(noSuchUser);
-    recipients.add(badRecipient1);
-    User badRecipient2 = getInvalidUser(wrongUser);
-    recipients.add(badRecipient2);
-    User goodRecipient = MMX.getCurrentUser();
-    recipients.add(goodRecipient);
-
-    HashMap<String, String> content = new HashMap<String, String>();
-    content.put("foo", "bar");
-    MMXMessage message = new MMXMessage.Builder()
-            .recipients(recipients)
-            .content(content)
-            .build();
-    final ExecMonitor<Boolean, MMXMessage.FailureCode> sendResult =
-            new ExecMonitor<Boolean, MMXMessage.FailureCode>();
-    final String messageId = message.send(new MMXMessage.OnFinishedListener<String>() {
-      @Override
-      public void onSuccess(String msgId) {
-        sendResult.invoked(Boolean.TRUE);
-      }
-      @Override
-      public void onFailure(MMXMessage.FailureCode code, Throwable ex) {
-        Log.d(TAG, "Multicast onFailure() ex="+ex);
-        if (code.equals(MMXMessage.FailureCode.INVALID_RECIPIENT)) {
-          InvalidRecipientException irEx = (InvalidRecipientException) ex;
-          invalidUsers.addAll(irEx.getUserIds());
-        }
-        sendResult.failed(code);
-      }
-    });
-    
-    // Send partial failure despite invalid recipients
-    ExecMonitor.Status status = sendResult.waitFor(TIMEOUT_IN_MILISEC);
-    assertEquals(ExecMonitor.Status.FAILED, status);
-    assertEquals(MMXMessage.FailureCode.INVALID_RECIPIENT, sendResult.getFailedValue());
-    assertEquals(2, invalidUsers.size());
-    assertTrue(invalidUsers.contains(badRecipient1.getUserIdentifier()));
-    assertTrue(invalidUsers.contains(badRecipient2.getUserIdentifier()));
-    
-    // Do a sleep; one error message per invalid recipient will be received.
-    try {
-      Thread.sleep(2000);
-    } catch (InterruptedException e) {
-      // Ignored.
-    }
-    // One message should arrive.
-    status = receivedResult.waitFor(1000);
-    assertEquals(ExecMonitor.Status.INVOKED, status);
-    assertTrue(receivedResult.getReturnValue());
-    
-    MMX.unregisterListener(messageListener);
-//    assertFalse(MMX.getMMXClient().isConnected());
-  }
 }
