+++ android/src/androidTest/java/com/magnet/mmx/client/utils/MaxHelper.java
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) 2012-2016 Magnet Systems. All rights reserved.
+ */
+package com.magnet.mmx.client.utils;
+
+import android.content.Context;
+import android.support.test.InstrumentationRegistry;
+import com.magnet.max.android.MaxCore;
+import com.magnet.mmx.client.api.MMX;
+
+public class MaxHelper {
+
+  public static void initMax() {
+    Context context = InstrumentationRegistry.getInstrumentation().getTargetContext();
+
+    MaxCore.init(context, new MaxAndroidJsonConfig(context, com.magnet.mmx.test.R.raw.keys));
+    MaxCore.register(MMX.getModule());
+
+    com.magnet.mmx.client.common.Log.setLoggable(null, com.magnet.mmx.client.common.Log.VERBOSE);
+  }
+
+}
