+++ android/src/androidTest/java/com/magnet/mmx/client/utils/TestConstants.java
@@ -1,0 +1,14 @@
+package com.magnet.mmx.client.utils;
+
+/**
+ * Copyright (c) 2012-2016 Magnet Systems. All rights reserved.
+ */
+public interface TestConstants {
+  int TIMEOUT_IN_SECOND = 5;
+  int TIMEOUT_IN_MILISEC = TIMEOUT_IN_SECOND * 1000;
+
+  //int LONG_TIMEOUT_IN_SECOND = 10;
+  //int LONG_TIMEOUT_IN_MILISEC = LONG_TIMEOUT_IN_SECOND * 1000;
+
+  int SLEEP_IN_MILISEC = 3 * 1000;
+}
